chore(deps): update hexo-generator-index to v4.0.0
This commit updates the `hexo-generator-index` package from version 3.0.0 to 4.0.0,
along with associated dependencies like `dompurify`, `domutils`, and `highlight.js`.
Changes are reflected in both `package-lock.json` and `package.json` to ensure the
latest versions are used for compatibility and security purposes. The update also
involves minor adjustments to the `package.json` file to accommodate the new version
requirements.